0.2.1 update (#2)
Functionality Changes and Feature adds
* `current_password` is now required for destroying a user record
* Initializer now lets users set some values instead of being hardcoded
* Install generators make getting up and running easier

Testing
* Converted from test unit to RSpec
* Mirror install setup into dummy app
* Tests now cover much of the functionality. Of note though, is a large TODO block on `/auth/validate` and /auth/passwords/reset`

Documentation
* Update and expand Readme
* Add Contributing
* Add Code of Conduct

Refactoring
* Password model is refactored and the controller changes made to match
* User mailers are made more generic (this should still be considered WIP)
* Gemspec dependencies now have version numbers

Bug fixes
* `users#update` now looks up users by ID instead of using the `current_user` helper (but it still uses it to get the id). When only using the helper, server would respond with whatever "updated" values were passed to it regardless of whether the record was saved.

+++ app/controllers/knock_once/users_controller.rb
@@ -1,4 +1,4 @@
-require_dependency "knock_once/application_controller"
+require_dependency 'knock_once/application_controller'
 
 module KnockOnce
   class UsersController < ApplicationController
@@ -11,7 +11,7 @@ module KnockOnce
     end
 
     def update
-      @user = current_user
+      @user = User.find_by_id(current_user.id)
       if @user.authenticate(params[:current_password])
         if @user.update(user_params)
           render json: {
@@ -28,10 +28,14 @@ module KnockOnce
 
     def destroy
       @user = current_user
-      if @user.destroy
-        render json: :success
+      if @user.authenticate(params[:current_password])
+        if @user.destroy
+          render json: :success
+        else
+          render json: @user.errors.full_messages
+        end
       else
-        render json: @user.errors.full_messages
+        render status: :unprocessable_entity, json: ['Current password is incorrect']
       end
     end
 
@@ -47,7 +51,7 @@ module KnockOnce
     private
 
     def user_params
-      params.permit(:user, :user_name, :email, :current_password, :password, :password_confirmation)
+      params.permit(KnockOnce.configuration.user_params)
     end
   end
 end
